move live-result sort to service
